pyproject.toml
[tool.poetry]
name = "pytorch-sample"
version = "0.1.0"
description = ""
authors = ["ariririri"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.5"
torchvision = "^0.6"
matplotlib = "^3.2"
Pillow = "^7.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
